进一步实现点赞评论模块，之后实现flink cdc异步更新redis

+++ mall-portal/src/main/java/com/macro/mall/portal/domain/CommentParam.java
@@ -8,5 +8,5 @@ public class CommentParam {
     private String content;
     private Long goodId;
     private String picUrl;
-    private int parentId;
+    private Long parentId;
 }
